+++ cocos/editor-support/cocostudio/CCSGUIReader.cpp
@@ -42,6 +42,7 @@ THE SOFTWARE.
 
 using namespace cocos2d::ui;
 using namespace cocos2d;
+using namespace cocos2d::layout;
 
 
 namespace cocostudio {
