+++ cocos/layout/CCLayoutParameterProtocol.cpp
@@ -1,35 +1,48 @@
-/****************************************************************************
- Copyright (c) 2013-2014 Chukong Technologies Inc.
- 
- http://www.cocos2d-x.org
- 
- Permission is hereby granted, free of charge, to any person obtaining a copy
- of this software and associated documentation files (the "Software"), to deal
- in the Software without restriction, including without limitation the rights
- to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the Software is
- furnished to do so, subject to the following conditions:
- 
- The above copyright notice and this permission notice shall be included in
- all copies or substantial portions of the Software.
- 
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- THE SOFTWARE.
- ****************************************************************************/
+//
+//  CCLayoutParameterProtocol.cpp
+//  cocos2d_libs
+//
+//  Created by cai wenzhi on 14-5-7.
+//
+//
 
-#include "ui/UILayoutParameter.h"
-#include "ui/UILayout.h"
+#include "layout/CCLayoutParameterProtocol.h"
 
 NS_CC_BEGIN
+namespace layout{
+    
+Margin::Margin(void) : left(0), top(0), right(0), bottom(0)
+{
+}
 
-namespace ui {
+Margin::Margin(float l, float t, float r, float b) : left(l), top(t), right(r), bottom(b)
+{
+}
 
+Margin::Margin(const Margin& other) : left(other.left), top(other.top), right(other.right), bottom(other.bottom)
+{
+}
 
+Margin& Margin::operator= (const Margin& other)
+{
+    setMargin(other.left, other.top, other.right, other.bottom);
+    return *this;
+}
+
+void Margin::setMargin(float l, float t, float r, float b)
+{
+    left = l;
+    top = t;
+    right = r;
+    bottom = b;
+}
+
+bool Margin::equals(const Margin &target) const
+{
+    return (left == target.left && top == target.top && right == target.right && bottom == target.bottom);
+}
+
+    
 LayoutParameter* LayoutParameter::create()
 {
     LayoutParameter* parameter = new LayoutParameter();
@@ -56,19 +69,19 @@ LayoutParameterType LayoutParameter::getLayoutType() const
 {
     return _layoutParameterType;
 }
-    
+
 LayoutParameter* LayoutParameter::clone()
 {
     LayoutParameter* clonedParameter = createCloneInstance();
     clonedParameter->copyProperties(this);
     return clonedParameter;
 }
-    
+
 LayoutParameter* LayoutParameter::createCloneInstance()
 {
     return LayoutParameter::create();
 }
-    
+
 void LayoutParameter::copyProperties(LayoutParameter *model)
 {
     _margin = model->_margin;
@@ -95,7 +108,7 @@ LinearGravity LinearLayoutParameter::getGravity() const
 {
     return _linearGravity;
 }
-    
+
 LayoutParameter* LinearLayoutParameter::createCloneInstance()
 {
     return LinearLayoutParameter::create();
@@ -152,7 +165,7 @@ const char* RelativeLayoutParameter::getRelativeName() const
 {
     return _relativeLayoutName.c_str();
 }
-    
+
 LayoutParameter* RelativeLayoutParameter::createCloneInstance()
 {
     return RelativeLayoutParameter::create();
@@ -169,7 +182,30 @@ void RelativeLayoutParameter::copyProperties(LayoutParameter *model)
         setRelativeToWidgetName(parameter->_relativeWidgetName.c_str());
     }
 }
-
+    
+LayoutParameterProtocol::LayoutParameterProtocol()
+{
+    
+}
+    
+LayoutParameterProtocol::~LayoutParameterProtocol()
+{
+    
+}
+    
+void LayoutParameterProtocol::setLayoutParameter(LayoutParameter *parameter)
+{
+    if (!parameter)
+    {
+        return;
+    }
+    _layoutParameterDictionary.insert(parameter->getLayoutType(), parameter);
 }
 
+LayoutParameter* LayoutParameterProtocol::getLayoutParameter(LayoutParameterType type)
+{
+    return dynamic_cast<LayoutParameter*>(_layoutParameterDictionary.at(type));
+}
+    
+}
 NS_CC_END
