+++ cocos/ui/UILayout.h
@@ -26,6 +26,7 @@ THE SOFTWARE.
 #define __LAYOUT_H__
 
 #include "ui/UIWidget.h"
+#include "layout/CCLayoutProtocol.h"
 
 NS_CC_BEGIN
 
@@ -38,26 +39,13 @@ typedef enum
     LAYOUT_COLOR_GRADIENT
 }LayoutBackGroundColorType;
 
-typedef enum
-{
-    LAYOUT_ABSOLUTE,
-    LAYOUT_LINEAR_VERTICAL,
-    LAYOUT_LINEAR_HORIZONTAL,
-    LAYOUT_RELATIVE
-}LayoutType;
-
 typedef enum {
     LAYOUT_CLIPPING_STENCIL,
     LAYOUT_CLIPPING_SCISSOR
 }LayoutClippingType;
 
-/**
- *  @js NA
- *  @lua NA
- */
-class LayoutExecutant;
     
-class Layout : public Widget
+class Layout : public Widget, public layout::LayoutProtocol
 {
     
     DECLARE_CLASS_GUI_INFO
@@ -208,7 +196,7 @@ public:
      *
      * @param LayoutType
      */
-    virtual void setLayoutType(LayoutType type);
+    virtual void setLayoutType(layout::LayoutType type);
     
     /**
      * Gets LayoutType.
@@ -217,7 +205,7 @@ public:
      *
      * @return LayoutType
      */
-    virtual LayoutType getLayoutType() const;
+    virtual layout::LayoutType getLayoutType() const;
 
     virtual void addChild(Node * child) override;
     /**
@@ -315,7 +303,7 @@ protected:
     //init background image renderer.
     void addBackGroundImage();
     
-    void supplyTheLayoutParameterLackToChild(Widget* child);
+    void supplyTheLayoutParameterLackToChild(LayoutParameterProtocol* child);
     virtual Widget* createCloneInstance() override;
     virtual void copySpecialProperties(Widget* model) override;
     virtual void copyClonedWidgetChildren(Widget* model) override;
@@ -337,7 +325,6 @@ protected:
     void updateBackGroundImageColor();
     void updateBackGroundImageOpacity();
     void updateBackGroundImageRGBA();
-    LayoutExecutant* createCurrentLayoutExecutant();
     
     /**
      *get the content size of the layout, it will accumulate all its children's content size
@@ -460,7 +447,6 @@ protected:
     Vector2 _alongVector;
     GLubyte _cOpacity;
     Size _backGroundImageTextureSize;
-    LayoutType _layoutType;
     LayoutClippingType _clippingType;
     DrawNode* _clippingStencil;
     bool _scissorRectDirty;
@@ -489,7 +475,6 @@ protected:
     Color3B _backGroundImageColor;
     GLubyte _backGroundImageOpacity;
     
-    LayoutExecutant* _curLayoutExecutant;
     
     GLint _mask_layer_le;
     GroupCommand _groupCommand;
