+++ cocos/ui/UIRelativeBox.cpp
@@ -64,7 +64,7 @@ bool RelativeBox::init()
 {
     if (Layout::init())
     {
-        setLayoutType(LAYOUT_RELATIVE);
+        setLayoutType(layout::LAYOUT_RELATIVE);
         return true;
     }
     return false;
