+++ tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UILayoutTest/UILayoutTest.cpp
@@ -5,6 +5,8 @@
 
 // UILayoutTest
 
+using namespace cocos2d::layout;
+
 UILayoutTest::UILayoutTest()
 {
 }
